+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Skip codegen when detecting descriptors won't change
+
 ## 0.17.1 - 2025-12-10
 
 ### Fixed